refactor: extract explosive and Ebb roll orchestration from actor sheet
Move explosive throw dialog, canvas targeting, deviation, and blast
region placement to explosive-rolls.mjs. Move Ebb formula roll flow to
ebb-rolls.mjs. Add pure explosive/Ebb helpers to roll-math.mjs with
unit tests. Actor sheet delegates via thin wrappers (~855 lines removed).
