+++ module/sheets/actor-sheet.mjs
@@ -5,12 +5,10 @@ import { LuckDialog } from '../apps/luck-dialog.mjs';
 import { XPDialog } from '../apps/xp-dialog.mjs';
 import { SlaSimpleContentDialog } from '../apps/sla-simple-dialog.mjs';
 import { calculateRollResult, generateDiceTooltip, createSLARoll } from '../helpers/dice.mjs';
-import { prepareItems, normalizeEbbEffect, normalizeEbbHealWoundMode, incrementSkillRank } from '../helpers/items.mjs';
-import { getEbbMosDamageBonus } from '../helpers/ebb-mos.mjs';
+import { prepareItems, incrementSkillRank } from '../helpers/items.mjs';
 import { applyMeleeModifiers, applyRangedModifiers, calculateRangePenalty } from '../helpers/modifiers.mjs';
 import { addActorItemToHotbar } from '../helpers/sla-hotbar.mjs';
 import { SLAChat } from '../helpers/chat.mjs';
-import { syncEbbCriticalFlux } from '../helpers/ebb-flux.mjs';
 import { handleStackableActorItemDrop } from '../helpers/inventory-stack.mjs';
 import {
     applySuccessThroughExperience,
@@ -20,6 +18,8 @@ import {
     computeSuccessDieOutcome
 } from './actor/roll-math.mjs';
 import { executeSkillRollFromItem } from './actor/skill-rolls.mjs';
+import { executeEbbRoll } from './actor/ebb-rolls.mjs';
+import { processExplosiveRoll, renderExplosiveDialog } from './actor/explosive-rolls.mjs';
 import { processWeaponRoll, renderAttackDialog } from './actor/weapon-rolls.mjs';
 
 const { HandlebarsApplicationMixin } = foundry.applications.api;
@@ -1344,871 +1344,16 @@ export class SlaActorSheet extends HandlebarsApplicationMixin(ActorSheetV2) {
         return processWeaponRoll(this, item, html, isMelee);
     }
 
-    // --- EXPLOSIVE LOGIC ---
     async _renderExplosiveDialog(item) {
-        // Prepare Template Data (Simplified for Explosives)
-        const templateData = {
-            item: item,
-            isMelee: false, // It's ranged/thrown usually
-            validModes: { single: { label: 'Single', active: true, rounds: 1, recoil: 0 } }, // Dummy structure
-            selectedMode: 'single',
-            recoil: 0
-        };
-
-        const content = await foundry.applications.handlebars.renderTemplate(
-            'systems/sla-industries/templates/dialogs/attack-dialog.hbs',
-            templateData
-        );
-
-        await new SlaSimpleContentDialog({
-            title: `Throw: ${item.name}`,
-            contentHtml: content,
-            width: 520,
-            classes: ['sla-dialog-window', 'dialog'],
-            actionLabel: 'THROW',
-            onConfirm: (root) => void this._processExplosiveRoll(item, root)
-        }).render(true);
-    }
-
-    _readExplosiveRollForm(form) {
-        return {
-            mod: Number(form.modifier?.value) || 0,
-            cover: Number(form.cover?.value) || 0,
-            aiming: form.aiming?.value || 'none',
-            blind: form.blind?.checked || false
-        };
-    }
-
-    _resolveExplosiveBlastData(item) {
-        const innerDist = item.system.blastRadiusInner || 0;
-        let outerDist = item.system.blastRadiusOuter || 0;
-        if (outerDist === 0) outerDist = 5;
-        return { innerDist, outerDist };
-    }
-
-    _resolveExplosiveSkillContext(item) {
-        const skillName = item.system.skill || 'throw';
-        const combatSkills = CONFIG.SLA?.combatSkills || {};
-        const resolvedSkillName = combatSkills[skillName] || skillName;
-
-        let rank = 0;
-        let skillItemForStat = null;
-        if (resolvedSkillName) {
-            const skillItem = this.actor.items.find(
-                (i) => i.type === 'skill' && i.name.trim().toLowerCase() === resolvedSkillName.trim().toLowerCase()
-            );
-            if (skillItem) {
-                rank = Number(skillItem.system.rank) || 0;
-                skillItemForStat = skillItem;
-            }
-        }
-
-        const statKey = skillItemForStat?.system?.stat || 'dex';
-        const statValue = this.actor.system.stats[statKey]?.total ?? this.actor.system.stats[statKey]?.value ?? 0;
-        const strValue = this.actor.system.stats.str?.total ?? this.actor.system.stats.str?.value ?? 0;
-        return { rank, statValue, strValue };
-    }
-
-    _buildExplosiveMods(rollData) {
-        return {
-            successDie: 0,
-            allDice: rollData.mod,
-            rank: 0,
-            damage: 0,
-            autoSkillSuccesses: 0
-        };
-    }
-
-    _applyExplosiveRollAdjustments(rollData, mods) {
-        if (this.actor.system.conditions?.prone) mods.allDice -= 1;
-        if (this.actor.system.conditions?.stunned) mods.allDice -= 1;
-        const penalty = this.actor.system.wounds.penalty || 0;
-        if (game.settings.get('sla-industries', 'enableAutomaticWoundPenalties')) {
-            mods.allDice -= penalty;
-        }
-
-        mods.successDie += rollData.cover;
-        if (rollData.aiming === 'sd') mods.successDie += 1;
-        if (rollData.aiming === 'skill') mods.autoSkillSuccesses += 1;
-    }
-
-    _getActorTokenForExplosiveRange() {
-        return this.token?.object ?? this.actor.getActiveTokens()[0];
-    }
-
-    _appendExplosiveRangeNotes(notes, item, strValue, target, token) {
-        if (item.system.blastRadiusInner || item.system.blastRadiusOuter) {
-            const txt =
-                item.system.blastRadiusInner > 0
-                    ? `${item.system.blastRadiusInner}/${item.system.blastRadiusOuter}m`
-                    : `${item.system.blastRadiusOuter}m`;
-            notes.push(`<strong>Blast:</strong> ${txt}`);
-        }
-
-        const effectiveRange = 15 + Math.min(Math.max(0, strValue), 5) * 5;
-        notes.push(`<strong>Max Range:</strong> ${effectiveRange}m`);
-
-        if (!token || target == null) return;
-        const ray = new foundry.canvas.geometry.Ray(token.center, target);
-        const distMeters = (ray.distance / canvas.scene.grid.size) * canvas.scene.grid.distance;
-        if (distMeters > effectiveRange) {
-            notes.push(`<strong style='color:#ffa500'>OUT OF RANGE (${Math.round(distMeters)}m)</strong>`);
-        }
-    }
-
-    _resolveDeviationWallCollision(start, end, epsilon = 2, source = null) {
-        const fallback = { x: end.x, y: end.y, blocked: false };
-        if (!canvas?.walls) return fallback;
-
-        const ray = new foundry.canvas.geometry.Ray(start, end);
-        const distanceFromStart = (point) => {
-            if (point?.x == null || point?.y == null) return Number.POSITIVE_INFINITY;
-            return Math.hypot(point.x - start.x, point.y - start.y);
-        };
-        const extractImpact = (collision) => {
-            if (!collision) return null;
-            const impact = collision.intersection || collision.point || collision;
-            if (impact?.x == null || impact?.y == null) return null;
-            return impact;
-        };
-        const resolveImpact = (collision) => {
-            if (!collision) return null;
-            if (Array.isArray(collision)) {
-                const impacts = collision.map(extractImpact).filter((p) => p?.x != null && p?.y != null);
-                if (!impacts.length) return null;
-                impacts.sort((a, b) => distanceFromStart(a) - distanceFromStart(b));
-                return resolveImpact(impacts[0]);
-            }
-
-            const impact = extractImpact(collision);
-            if (impact?.x == null || impact?.y == null) return null;
-
-            // Pull back slightly so the template center never crosses the wall boundary.
-            const impactDistance = distanceFromStart(impact);
-            if (impactDistance <= epsilon) return { x: start.x, y: start.y, blocked: true };
-            const safeDistance = Math.max(0, impactDistance - epsilon);
-            const t = safeDistance / (impactDistance || 1);
-            return {
-                x: start.x + (impact.x - start.x) * t,
-                y: start.y + (impact.y - start.y) * t,
-                blocked: true
-            };
-        };
-
-        try {
-            const backend = CONFIG?.Canvas?.polygonBackends?.move;
-            const backendClosest = backend?.testCollision
-                ? backend.testCollision(start, end, { type: 'move', mode: 'closest', source })
-                : null;
-            const resolvedBackendClosest = resolveImpact(backendClosest);
-            if (resolvedBackendClosest) return resolvedBackendClosest;
-        } catch (_err) {
-            // Keep fallback path active for API/version differences.
-        }
-
-        try {
-            const closest = canvas.walls.checkCollision(ray, { type: 'move', mode: 'closest', source });
-            const resolvedClosest = resolveImpact(closest);
-            if (resolvedClosest) return resolvedClosest;
-        } catch (_err) {
-            // Keep fallback path active for API/version differences.
-        }
-
-        try {
-            const any = canvas.walls.checkCollision(ray, { type: 'move', mode: 'any', source });
-            if (any === true) {
-                const collisions = canvas.walls.checkCollision(ray, { type: 'move', mode: 'all', source });
-                const resolvedAll = resolveImpact(collisions);
-                if (resolvedAll) return resolvedAll;
-            }
-        } catch (_err) {
-            // Keep fallback path active for API/version differences.
-        }
-
-        return fallback;
-    }
-
-    _resolveExplosiveDeviation({ isBaseSuccess, skillSuccessCount, target, token }) {
-        const noCanvasTarget = target == null;
-        let outcomeText = '';
-        let resultColor = '#f55';
-        let isSuccess = false;
-        let finalX = noCanvasTarget ? 0 : target.x;
-        let finalY = noCanvasTarget ? 0 : target.y;
-        let wallBlocked = false;
-
-        const allDiceFailed = !isBaseSuccess && skillSuccessCount === 0;
-        if (allDiceFailed) {
-            outcomeText = "<strong style='color:#ff0000; font-size:1.1em;'>FUMBLE: Detonates on Thrower!</strong>";
-            resultColor = '#ff0000';
-            if (token) {
-                finalX = token.center.x;
-                finalY = token.center.y;
-            }
-            return { outcomeText, resultColor, isSuccess, finalX, finalY, wallBlocked };
-        }
-
-        if (isBaseSuccess && skillSuccessCount > 0) {
-            outcomeText = noCanvasTarget
-                ? "<strong style='color:#39ff14'>SUCCESS</strong>"
-                : "<strong style='color:#39ff14'>LANDS ON TARGET</strong>";
-            resultColor = '#39ff14';
-            isSuccess = true;
-            if (noCanvasTarget) {
-                finalX = 0;
-                finalY = 0;
-            }
-            return { outcomeText, resultColor, isSuccess, finalX, finalY, wallBlocked };
-        }
-
-        const devMeters = isBaseSuccess ? 5 : 10;
-        outcomeText = isBaseSuccess
-            ? "<strong style='color:#ffa500'>DEVIATION: 5m</strong>"
-            : "<strong style='color:#ff5555'>DEVIATION: 10m</strong>";
-        resultColor = isBaseSuccess ? '#ffa500' : '#ff5555';
-
-        if (noCanvasTarget) {
-            const originX = token?.center?.x ?? 0;
-            const originY = token?.center?.y ?? 0;
-            const grid = canvas?.scene?.grid;
-            if (grid?.distance && grid?.size) {
-                const devPixels = (devMeters / grid.distance) * grid.size;
-                const angle = Math.random() * 2 * Math.PI;
-                finalX = originX + Math.cos(angle) * devPixels;
-                finalY = originY + Math.sin(angle) * devPixels;
-            } else {
-                finalX = originX;
-                finalY = originY;
-            }
-            return { outcomeText, resultColor, isSuccess, finalX, finalY, wallBlocked: false };
-        }
-
-        const devPixels = (devMeters / canvas.scene.grid.distance) * canvas.scene.grid.size;
-        const angle = Math.random() * 2 * Math.PI;
-        finalX = target.x + Math.cos(angle) * devPixels;
-        finalY = target.y + Math.sin(angle) * devPixels;
-
-        const wallCollision = this._resolveDeviationWallCollision(
-            target,
-            { x: finalX, y: finalY },
-            2,
-            token?.document ?? null
-        );
-        finalX = wallCollision.x;
-        finalY = wallCollision.y;
-        wallBlocked = wallCollision.blocked;
-
-        return { outcomeText, resultColor, isSuccess, finalX, finalY, wallBlocked };
-    }
-
-    async _placeExplosiveTemplates({ item, blastRadius, innerDist, finalX, finalY, isSuccess }) {
-        try {
-            const distancePixels =
-                canvas?.dimensions?.distancePixels ?? canvas.scene.grid.size / canvas.scene.grid.distance;
-            const visMode = game.settings.get('sla-industries', 'blastRegionVisibility') ?? 'observer';
-            const visibility = visMode === 'always' ? CONST.REGION_VISIBILITY.ALWAYS : CONST.REGION_VISIBILITY.OBSERVER;
-            const baseRegionData = {
-                color: game.user.color,
-                displayMeasurements: true,
-                visibility,
-                ownership: { [game.user.id]: CONST.DOCUMENT_OWNERSHIP_LEVELS.OWNER },
-                ...(canvas.level?.id ? { levels: [canvas.level.id] } : {})
-            };
-            const regions = [
-                {
-                    ...baseRegionData,
-                    name: `${item.name} (Outer)`,
-                    flags: { sla: { itemId: item.id, isDeviation: !isSuccess, type: 'outer' } },
-                    shapes: [
-                        {
-                            type: 'circle',
-                            x: finalX,
-                            y: finalY,
-                            radius: blastRadius * distancePixels,
-                            gridBased: false
-                        }
-                    ]
-                }
-            ];
-
-            if (innerDist > 0 && innerDist < blastRadius) {
-                regions.push({
-                    ...baseRegionData,
-                    name: `${item.name} (Inner)`,
-                    flags: { sla: { itemId: item.id, isDeviation: !isSuccess, type: 'inner' } },
-                    shapes: [
-                        {
-                            type: 'circle',
-                            x: finalX,
-                            y: finalY,
-                            radius: innerDist * distancePixels,
-                            gridBased: false
-                        }
-                    ]
-                });
-            }
-
-            await canvas.scene.createEmbeddedDocuments('Region', regions);
-        } catch (err) {
-            console.error('SLA | Template Creation Failed:', err);
-        }
+        return renderExplosiveDialog(this, item);
     }
 
     async _processExplosiveRoll(item, html) {
-        const root = html?.jquery ? html[0] : html;
-        const form = root instanceof HTMLFormElement ? root : root?.querySelector?.('form');
-        if (!form) return;
-
-        const rollData = this._readExplosiveRollForm(form);
-        const { innerDist, outerDist } = this._resolveExplosiveBlastData(item);
-        const automateThrow = game.settings.get('sla-industries', 'enableExplosiveThrowAutomation');
-        let target = null;
-        if (automateThrow) {
-            ui.notifications.info('Select target position...');
-            target = await this._waitForCanvasClick({ outerDist, innerDist });
-            if (!target) return;
-        }
-        await this._resolveExplosiveRoll(item, rollData, target, outerDist, innerDist);
-    }
-
-    _drawDashedCircle(graphics, centerX, centerY, radius, color, thickness, dashLength, gapLength) {
-        if (!graphics || radius <= 0) return;
-        const circumference = Math.PI * 2 * radius;
-        const segmentLength = Math.max(1, dashLength + gapLength);
-        const segmentCount = Math.max(12, Math.floor(circumference / segmentLength));
-        const angleStep = (Math.PI * 2) / segmentCount;
-        const dashStep = Math.max(1, Math.floor(segmentCount * (dashLength / segmentLength)));
-
-        graphics.lineStyle(thickness, color, 1);
-        for (let i = 0; i < segmentCount; i += dashStep * 2) {
-            for (let j = 0; j < dashStep; j++) {
-                const idx = i + j;
-                if (idx >= segmentCount) break;
-                const a1 = idx * angleStep;
-                const a2 = (idx + 1) * angleStep;
-                const x1 = centerX + Math.cos(a1) * radius;
-                const y1 = centerY + Math.sin(a1) * radius;
-                const x2 = centerX + Math.cos(a2) * radius;
-                const y2 = centerY + Math.sin(a2) * radius;
-                graphics.moveTo(x1, y1);
-                graphics.lineTo(x2, y2);
-            }
-        }
-    }
-
-    _waitForCanvasClick({ outerDist = 0, innerDist = 0 } = {}) {
-        return new Promise((resolve) => {
-            const stage = canvas?.app?.stage;
-            if (!stage) {
-                resolve(null);
-                return;
-            }
-
-            const distancePixels =
-                canvas?.dimensions?.distancePixels ?? canvas?.scene?.grid?.size / canvas?.scene?.grid?.distance;
-            const outerRadiusPx = Math.max(0, Number(outerDist) || 0) * (distancePixels || 0);
-            const innerRadiusPx = Math.max(0, Number(innerDist) || 0) * (distancePixels || 0);
-            const preview = new PIXI.Graphics();
-            preview.eventMode = 'none';
-            preview.zIndex = 9999;
-            stage.addChild(preview);
-
-            const redrawPreview = (x, y) => {
-                preview.clear();
-
-                // Faint fill + dashed edge for target preview.
-                if (outerRadiusPx > 0) {
-                    preview.beginFill(0xffaa00, 0.08);
-                    preview.drawCircle(x, y, outerRadiusPx);
-                    preview.endFill();
-                    this._drawDashedCircle(preview, x, y, outerRadiusPx, 0xffaa00, 2, 7, 6);
-                }
-
-                if (innerRadiusPx > 0 && innerRadiusPx < outerRadiusPx) {
-                    preview.beginFill(0xff4444, 0.1);
-                    preview.drawCircle(x, y, innerRadiusPx);
-                    preview.endFill();
-                    this._drawDashedCircle(preview, x, y, innerRadiusPx, 0xff4444, 2, 5, 5);
-                }
-            };
-
-            const cleanup = () => {
-                stage.off('click', clickHandler);
-                stage.off('pointermove', moveHandler);
-                stage.off('rightdown', cancelHandler);
-                if (preview.parent) preview.parent.removeChild(preview);
-                preview.destroy();
-            };
-
-            const moveHandler = (event) => {
-                const pos = event?.data?.getLocalPosition(stage);
-                if (!pos) return;
-                redrawPreview(pos.x, pos.y);
-            };
-
-            // Initialize preview at current mouse position when possible.
-            if (canvas?.mousePosition) {
-                redrawPreview(canvas.mousePosition.x, canvas.mousePosition.y);
-            }
-
-            stage.on('pointermove', moveHandler);
-
-            const clickHandler = (event) => {
-                event.stopPropagation();
-                // Get world coords
-                const pos = event.data.getLocalPosition(stage);
-                cleanup();
-                resolve({ x: pos.x, y: pos.y });
-            };
-            stage.on('click', clickHandler);
-
-            // Allow cancelling with Right Click
-            const cancelHandler = (event) => {
-                event?.stopPropagation?.();
-                cleanup();
-                resolve(null);
-            };
-            stage.on('rightdown', cancelHandler);
-        });
-    }
-
-    async _resolveExplosiveRoll(item, rollData, target, blastRadius, innerDist) {
-        const currentQty = item.system.quantity || 0;
-        if (currentQty <= 0) {
-            return ui.notifications.warn(`You are out of ${item.name}s.`);
-        }
-
-        const newQty = currentQty - 1;
-        if (newQty === 0) {
-            await item.delete();
-        } else {
-            await item.update({ 'system.quantity': newQty });
-        }
-
-        const { rank, statValue, strValue } = this._resolveExplosiveSkillContext(item);
-        const mods = this._buildExplosiveMods(rollData);
-        let notes = [];
-        const token = this._getActorTokenForExplosiveRange();
-        let resolvedTarget = target;
-        if (target != null && token) {
-            const throwCollision = this._resolveDeviationWallCollision(token.center, target, 2, token.document ?? null);
-            if (throwCollision.blocked) {
-                resolvedTarget = { x: throwCollision.x, y: throwCollision.y };
-                notes.push('<strong>Throw:</strong> Stopped by wall.');
-                ui.notifications.info(`${item.name} hit a wall before reaching the target point.`);
-            }
-        }
-
-        this._appendExplosiveRangeNotes(notes, item, strValue, resolvedTarget, token);
-        this._applyExplosiveRollAdjustments(rollData, mods);
-
-        const baseModifier = statValue + rank + mods.allDice;
-        const skillDiceCount = Math.max(0, rank + 1 + mods.rank);
-        const rollFormula = `1d10 + ${skillDiceCount}d10`;
-        let roll = createSLARoll(rollFormula);
-        await roll.evaluate();
-
-        const TN = 10;
-        const { sdTotal, isBaseSuccess } = this._computeSuccessDieOutcome({
-            roll,
-            baseModifier,
-            successDieModifier: mods.successDie,
-            targetNumber: TN
-        });
-
-        const { skillDiceData, skillSuccessCount } = this._buildSkillDiceResults({
-            roll,
-            baseModifier,
-            targetNumber: TN,
-            autoSuccesses: mods.autoSkillSuccesses
-        });
-
-        const deviationData = this._resolveExplosiveDeviation({
-            isBaseSuccess,
-            skillSuccessCount,
-            target: resolvedTarget,
-            token
-        });
-        const { outcomeText, resultColor, isSuccess, finalX, finalY, wallBlocked } = deviationData;
-
-        if (wallBlocked) {
-            notes.push('<strong>Deviation:</strong> Stopped by wall.');
-            ui.notifications.info(`${item.name} deviation hit a wall.`);
-        }
-
-        if (innerDist > 0) {
-            notes.push(`<br/><strong>Kill Zone (< ${innerDist}m):</strong> +2 Damage`);
-        }
-
-        if (game.settings.get('sla-industries', 'enableExplosiveThrowAutomation') && target != null) {
-            await this._placeExplosiveTemplates({ item, blastRadius, innerDist, finalX, finalY, isSuccess });
-        }
-
-        let baseDmg = item.system.damage || '0';
-        const adValue = Number(item.system.ad) || 0;
-
-        const notesText = notes.join(' ');
-        const templateData = {
-            actorUuid: this.actor.uuid,
-            borderColor: resultColor,
-            headerColor: resultColor,
-            resultColor: resultColor,
-            itemName: item.name.toUpperCase(),
-            successTotal: sdTotal,
-            tooltip: this._generateTooltip(roll, baseModifier, mods.successDie),
-            skillDice: skillDiceData,
-            notes: notesText,
-            showDamageButton: true,
-            dmgFormula: baseDmg,
-            dmgDisplay: this._resolveDamageDisplay(baseDmg),
-            minDamage: Number(item.system.minDamage) || 0,
-            adValue: adValue,
-            mos: {
-                isSuccess: isSuccess,
-                hits: skillSuccessCount,
-                effect: outcomeText
-            },
-            canUseLuck: this.actor.system.stats.luck.value > 0,
-            luckValue: this.actor.system.stats.luck.value,
-            isEbb: true // Legacy chat template flag for this non-weapon card path.
-        };
-
-        const chatContent = await foundry.applications.handlebars.renderTemplate(
-            'systems/sla-industries/templates/chat/chat-weapon-rolls.hbs',
-            templateData
-        );
-
-        roll.toMessage({
-            speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-            content: chatContent,
-            flags: {
-                sla: this._buildSlaRollFlags({
-                    baseModifier: baseModifier,
-                    itemName: item.name.toUpperCase(),
-                    notes: notesText,
-                    tn: 10,
-                    extra: {
-                        targets: Array.from(game.user.targets).map((t) => t.document.uuid),
-                        damageBase: baseDmg,
-                        adValue: adValue
-                    }
-                })
-            }
-        });
-    }
-
-    _resolveEbbDisciplineName(disciplineName) {
-        let resolvedName = disciplineName;
-        const ebbDisciplines = CONFIG.SLA?.ebbDisciplines || {};
-        for (const [key, label] of Object.entries(ebbDisciplines)) {
-            if (key === disciplineName || label === disciplineName) {
-                resolvedName = label;
-                break;
-            }
-        }
-        return resolvedName;
-    }
-
-    _resolveEbbContext(item) {
-        const formulaRating = item.system.formulaRating || 7;
-        const currentFlux = this.actor.system.stats.flux?.value || 0;
-        const fluxCost = Math.max(0, Number(item.system.cost) || 1);
-        const disciplineName = item.system.discipline;
-        const resolvedDisciplineName = this._resolveEbbDisciplineName(disciplineName);
-        const disciplineItem = this.actor.items.find(
-            (i) => i.type === 'discipline' && i.name.toLowerCase() === resolvedDisciplineName.toLowerCase()
-        );
-
-        return { formulaRating, currentFlux, fluxCost, resolvedDisciplineName, disciplineItem };
-    }
-
-    _calculateEbbModifier(rank) {
-        const statKey = 'conc';
-        const statValue = this.actor.system.stats[statKey]?.total ?? this.actor.system.stats[statKey]?.value ?? 0;
-        let globalMod = 0;
-        if (this.actor.system.conditions?.prone) globalMod -= 1;
-        if (this.actor.system.conditions?.stunned) globalMod -= 1;
-        const penalty = this.actor.system.wounds.penalty || 0;
-        const woundPenalty = game.settings.get('sla-industries', 'enableAutomaticWoundPenalties') ? penalty : 0;
-        return statValue + rank - woundPenalty + globalMod;
-    }
-
-    async _createAndEvaluateEbbRoll(rank) {
-        const skillDiceCount = rank + 1;
-        const rollFormula = `1d10 + ${skillDiceCount}d10`;
-        let roll = new Roll(rollFormula);
-
-        if (roll.terms.length > 0 && roll.terms[0].constructor.name === 'Die') {
-            roll.terms[0].options.appearance = {
-                foreground: '#FFFFFF',
-                background: '#000000',
-                edge: '#333333'
-            };
-        }
-
-        await roll.evaluate();
-        return roll;
-    }
-
-    _collectEbbSkillDiceData(roll, modifier, formulaRating) {
-        const { skillDiceData, skillSuccessCount } = this._buildSkillDiceResults({
-            roll,
-            baseModifier: modifier,
-            targetNumber: formulaRating
-        });
-        return { skillDiceData, skillSuccesses: skillSuccessCount };
-    }
-
-    _resolveEbbOutcomeText(isBaseSuccess, skillSuccesses, ebbEffectRaw) {
-        const allDiceFailed = !isBaseSuccess && skillSuccesses === 0;
-        // Formula succeeds only when the success die hits the rating; skill dice alone must not unlock damage/heal/wounds/effects.
-        const isSuccessful = isBaseSuccess;
-        const ebbEffect = normalizeEbbEffect(ebbEffectRaw);
-        const attackMos = ebbEffect === 'damage';
-
-        let mosEffectText = 'Standard Success';
-        let failureConsequence = 'Failed';
-
-        if (isSuccessful) {
-            if (skillSuccesses === 2) {
-                mosEffectText = attackMos ? '+1 Damage / Effect' : 'Standard Success';
-            } else if (skillSuccesses === 3) {
-                mosEffectText = attackMos
-                    ? '+2 Damage / Repeat Ability'
-                    : 'May use the same Ebb ability again within 5 minutes (-3 FLUX)';
-            } else if (skillSuccesses >= 4) {
-                mosEffectText = attackMos
-                    ? "<strong style='color:#39ff14'>CRITICAL:</strong> +4 Dmg | Regain 1 FLUX"
-                    : "<strong style='color:#39ff14'>CRITICAL:</strong> Regain 1 FLUX";
-            }
-        } else if (allDiceFailed) {
-            failureConsequence = "<strong style='color:#ff5555'>SEVERE FAILURE:</strong> -3 HP & -1 Extra FLUX";
-        }
-
-        return { isSuccessful, mosEffectText, failureConsequence };
-    }
-
-    _buildEbbDamageFormula(item, isSuccessful, skillSuccesses) {
-        const rawBase = item.system.dmg || item.system.damage || '0';
-        const baseDmg = String(rawBase);
-        const ebbEffect = normalizeEbbEffect(item.system.ebbEffect);
-        const mosDamageBonus = getEbbMosDamageBonus(isSuccessful, skillSuccesses, item.system.ebbEffect);
-
-        let finalDmgFormula = baseDmg;
-        if (baseDmg !== '0' && baseDmg !== '' && mosDamageBonus > 0) {
-            finalDmgFormula = `${baseDmg} + ${mosDamageBonus}`;
-        }
-
-        const hasHpFormula = finalDmgFormula && finalDmgFormula !== '0' && String(finalDmgFormula).trim() !== '';
-        const showHpRollButton = Boolean(
-            isSuccessful && (ebbEffect === 'damage' || ebbEffect === 'heal') && hasHpFormula
-        );
-        const removeWoundsCount = Math.max(0, Math.min(6, Math.floor(Number(item.system.removeWounds) || 0)));
-        const isHealEffect = ebbEffect === 'heal';
-        const healWoundMode = normalizeEbbHealWoundMode(item.system.ebbHealWoundMode);
-        const showRemoveWoundsOnly = Boolean(
-            isSuccessful &&
-            removeWoundsCount > 0 &&
-            ((ebbEffect === 'effect' && !hasHpFormula) || (isHealEffect && healWoundMode === 'or' && hasHpFormula))
-        );
-        const removeWoundsBundledWithHpRoll = Boolean(
-            showHpRollButton && removeWoundsCount > 0 && !(isHealEffect && healWoundMode === 'or')
-        )
-            ? removeWoundsCount
-            : 0;
-
-        return {
-            finalDmgFormula,
-            showDamageButton: showHpRollButton,
-            showHpRollButton,
-            showRemoveWoundsOnly,
-            removeWoundsBundledWithHpRoll,
-            healWoundMode,
-            ebbEffect,
-            isHealRoll: ebbEffect === 'heal'
-        };
-    }
-
-    _buildEbbTemplateData({
-        item,
-        roll,
-        modifier,
-        resultColor,
-        successTotal,
-        skillDiceData,
-        formulaRating,
-        showDamageButton,
-        showRemoveWoundsOnly,
-        removeWoundsBundledWithHpRoll,
-        healWoundMode,
-        finalDmgFormula,
-        isSuccessful,
-        skillSuccesses,
-        mosEffectText,
-        failureConsequence,
-        isHealRoll,
-        ebbEffect
-    }) {
-        const effectCount = item.effects?.size ?? 0;
-        const ebbTarget = item.system.ebbTarget || 'enemy';
-        const minDamageRaw = item.system.minDamage;
-        const minDamage =
-            minDamageRaw !== undefined && minDamageRaw !== null && String(minDamageRaw).trim() !== ''
-                ? minDamageRaw
-                : '0';
-        const i18n = (k) => game.i18n?.localize(k) ?? k;
-        const targetLabel = i18n(`SLA.EbbTarget.${ebbTarget}`);
-        const effectLabel = i18n(`SLA.EbbEffect.${ebbEffect || 'damage'}`);
-        const ebbHealWoundModeLabel = i18n(`SLA.EbbHealWoundMode.${healWoundMode === 'or' ? 'orHint' : 'andHint'}`);
-        const showEbbHealWoundModeHint = Boolean(
-            isHealRoll &&
-            showDamageButton &&
-            Math.max(0, Math.min(6, Math.floor(Number(item.system.removeWounds) || 0))) > 0
-        );
-        return {
-            borderColor: resultColor,
-            headerColor: resultColor,
-            resultColor: resultColor,
-            actorUuid: this.actor.uuid,
-            itemName: item.name.toUpperCase(),
-            successTotal: successTotal,
-            tooltip: this._generateTooltip(roll, modifier, 0),
-            skillDice: skillDiceData,
-            notes: `<strong>Formula Rating:</strong> ${formulaRating}`,
-            showDamageButton: showDamageButton,
-            showRemoveWoundsOnly: showRemoveWoundsOnly,
-            dmgFormula: finalDmgFormula,
-            dmgDisplay: this._resolveDamageDisplay(finalDmgFormula),
-            minDamage,
-            adValue: item.system.ad || 0,
-            mos: {
-                isSuccess: isSuccessful,
-                hits: skillSuccesses,
-                effect: isSuccessful ? mosEffectText : failureConsequence
-            },
-            isEbb: true,
-            ebbFormulaRoll: true,
-            isHealRoll: Boolean(isHealRoll),
-            ebbTarget,
-            ebbEffect: ebbEffect || 'damage',
-            ebbTargetLabel: targetLabel,
-            ebbEffectLabel: effectLabel,
-            removeWoundsCount: Math.max(0, Math.min(6, Math.floor(Number(item.system.removeWounds) || 0))),
-            removeWoundsBundledWithHpRoll,
-            ebbHealWoundMode: healWoundMode,
-            ebbHealWoundModeLabel,
-            showEbbHealWoundModeHint,
-            showEbbEffectButtons: Boolean(isSuccessful && effectCount > 0 && ebbTarget !== 'self'),
-            showEbbEffectSelfButton: Boolean(isSuccessful && effectCount > 0 && ebbTarget === 'self')
-        };
+        return processExplosiveRoll(this, item, html);
     }
 
     async _executeEbbRoll(item) {
-        const { formulaRating, currentFlux, fluxCost, resolvedDisciplineName, disciplineItem } =
-            this._resolveEbbContext(item);
-        if (currentFlux < fluxCost) {
-            ui.notifications.error('Insufficient FLUX.');
-            return;
-        }
-        await this.actor.update({ 'system.stats.flux.value': Math.max(0, currentFlux - fluxCost) });
-
-        if (!disciplineItem) {
-            ui.notifications.warn(`Missing Discipline Item: ${resolvedDisciplineName}`);
-            return;
-        }
-
-        const rank = Number(disciplineItem.system.rank) || 0;
-        const modifier = this._calculateEbbModifier(rank);
-        const roll = await this._createAndEvaluateEbbRoll(rank);
-        const { sdTotal: successTotal, isBaseSuccess } = this._computeSuccessDieOutcome({
-            roll,
-            baseModifier: modifier,
-            successDieModifier: 0,
-            targetNumber: formulaRating
-        });
-        const resultColor = isBaseSuccess ? '#39ff14' : '#f55';
-
-        const { skillDiceData, skillSuccesses } = this._collectEbbSkillDiceData(roll, modifier, formulaRating);
-        const { isSuccessful, mosEffectText, failureConsequence } = this._resolveEbbOutcomeText(
-            isBaseSuccess,
-            skillSuccesses,
-            item.system.ebbEffect
-        );
-        const {
-            finalDmgFormula,
-            showDamageButton,
-            showRemoveWoundsOnly,
-            removeWoundsBundledWithHpRoll,
-            healWoundMode,
-            ebbEffect,
-            isHealRoll
-        } = this._buildEbbDamageFormula(item, isSuccessful, skillSuccesses);
-        const notesText = `<strong>Formula Rating:</strong> ${formulaRating}`;
-        const templateData = this._buildEbbTemplateData({
-            item,
-            roll,
-            modifier,
-            resultColor,
-            successTotal,
-            skillDiceData,
-            formulaRating,
-            showDamageButton,
-            showRemoveWoundsOnly,
-            removeWoundsBundledWithHpRoll,
-            healWoundMode,
-            finalDmgFormula,
-            isSuccessful,
-            skillSuccesses,
-            mosEffectText,
-            failureConsequence,
-            isHealRoll,
-            ebbEffect
-        });
-
-        const chatContent = await foundry.applications.handlebars.renderTemplate(
-            'systems/sla-industries/templates/chat/chat-weapon-rolls.hbs',
-            templateData
-        );
-
-        const ebbEffectCount = item.effects?.size ?? 0;
-        const ebbRemoveWoundsCount = Math.max(0, Math.min(6, Math.floor(Number(item.system.removeWounds) || 0)));
-        const ebbHealWoundMutualExclude = Boolean(
-            isSuccessful &&
-            normalizeEbbEffect(item.system.ebbEffect) === 'heal' &&
-            normalizeEbbHealWoundMode(item.system.ebbHealWoundMode) === 'or' &&
-            ebbRemoveWoundsCount > 0 &&
-            showDamageButton
-        );
-        const message = await roll.toMessage({
-            speaker: ChatMessage.getSpeaker({ actor: this.actor }),
-            content: chatContent,
-            flags: {
-                sla: this._buildSlaRollFlags({
-                    baseModifier: modifier,
-                    itemName: item.name.toUpperCase(),
-                    notes: notesText,
-                    tn: formulaRating,
-                    extra: {
-                        isWeapon: false,
-                        isEbb: true,
-                        itemUuid: item.uuid,
-                        ebbHasEffects: ebbEffectCount > 0,
-                        ebbRollSuccess: isSuccessful,
-                        ebbTarget: item.system.ebbTarget || 'enemy',
-                        ebbEffect: normalizeEbbEffect(item.system.ebbEffect),
-                        actorUuid: this.actor.uuid,
-                        ebbRemoveWoundsCount: ebbRemoveWoundsCount,
-                        ebbHealWoundMode: normalizeEbbHealWoundMode(item.system.ebbHealWoundMode),
-                        ebbHealWoundMutualExclude,
-                        targets: Array.from(game.user.targets).map((t) => t.document.uuid)
-                    }
-                })
-            }
-        });
-        const chatMsg = Array.isArray(message) ? message[0] : message;
-        if (chatMsg) {
-            await syncEbbCriticalFlux(chatMsg, this.actor, chatMsg.flags?.sla ?? {}, isSuccessful, skillSuccesses);
-        }
+        return executeEbbRoll(this, item);
     }
 
     // --- DROP ITEM HANDLER ---
